Added Details and enhanced functionality
