+++ src/main/scala/caravan/bus/tilelink/StallUnit.scala
@@ -2,10 +2,12 @@ package caravan.bus.tileink
 
 import chisel3._ 
 
+// Module to Stall the Response one Cycle
+
 class stallUnit extends Module {
 
     implicit val config = TilelinkConfig()
-    
+
     val io = IO(new Bundle{
         val bundle_in = Input(new TilelinkSlave)
         val valid_in = Input(UInt(1.W))
