+++ src/main/scala/caravan/bus/tilelink/TilelinkDevice.scala
@@ -15,15 +15,18 @@ class TilelinkDevice(implicit val config: TilelinkConfig) extends DeviceAdapter 
     io.tlMasterReceiver.ready := true.B
     io.rspIn.ready := true.B
 
+    
     val stall = Module(new stallUnit)
 
+
+    // Sending Response coming from Memory in the STALL to delay the response one cycle
     stall.io.bundle_in.d_opcode := Mux(io.tlMasterReceiver.bits.a_opcode === Get.U, AccessAckData.U, AccessAck.U)
     stall.io.bundle_in.d_data := io.rspIn.bits.dataResponse
     stall.io.bundle_in.d_param := 0.U
     stall.io.bundle_in.d_size := io.tlMasterReceiver.bits.a_size
     stall.io.bundle_in.d_source := io.tlMasterReceiver.bits.a_source
     stall.io.bundle_in.d_sink := 0.U
-    stall.io.bundle_in.d_denied := 0.U
+    stall.io.bundle_in.d_denied := false.B
     stall.io.bundle_in.d_corrupt := io.rspIn.bits.error
     stall.io.valid_in := true.B
 
@@ -33,7 +36,7 @@ class TilelinkDevice(implicit val config: TilelinkConfig) extends DeviceAdapter 
     io.reqOut.bits.addrRequest := io.tlMasterReceiver.bits.a_address
     io.reqOut.bits.dataRequest := io.tlMasterReceiver.bits.a_data
     io.reqOut.bits.activeByteLane := io.tlMasterReceiver.bits.a_mask
-    io.reqOut.bits.isWrite := io.tlMasterReceiver.bits.a_opcode === PutFullData.U
+    io.reqOut.bits.isWrite := io.tlMasterReceiver.bits.a_opcode === PutFullData.U || io.tlMasterReceiver.bits.a_opcode === PutPartialData.U
     io.reqOut.valid := true.B
 
 }
