+++ src/main/scala/caravan/bus/tilelink/TilelinkHost.scala
@@ -3,7 +3,7 @@ import caravan.bus.common.HostAdapter
 import chisel3._
 import chisel3.experimental.DataMirror
 import chisel3.stage.ChiselStage
-import chisel3.util.{Decoupled, Enum, MuxCase}
+import chisel3.util.{Decoupled, Enum, MuxLookup}
 
 class TilelinkHost(implicit val config: TilelinkConfig) extends HostAdapter with OpCodes {
     val io = IO(new Bundle {
@@ -18,14 +18,19 @@ class TilelinkHost(implicit val config: TilelinkConfig) extends HostAdapter with
 
 
     
-    io.tlMasterTransmitter.bits.a_opcode := Mux(io.reqIn.bits.isWrite, PutFullData.U, Get.U)
+    io.tlMasterTransmitter.bits.a_opcode := Mux(io.reqIn.bits.isWrite, Mux(io.reqIn.bits.activeByteLane === "b1111".U, PutFullData.U, PutPartialData.U) , Get.U)
     io.tlMasterTransmitter.bits.a_data := io.reqIn.bits.dataRequest
     io.tlMasterTransmitter.bits.a_address := io.reqIn.bits.addrRequest
     io.tlMasterTransmitter.bits.a_param := 0.U
-    io.tlMasterTransmitter.bits.a_source := 2.U
-    io.tlMasterTransmitter.bits.a_size := 2.U
+    io.tlMasterTransmitter.bits.a_source := 2.U 
+    io.tlMasterTransmitter.bits.a_size := MuxLookup(config.w.U, 2.U,Array(                    // default 32-bit
+                                                                            (1.U) -> 0.U,
+                                                                            (2.U) -> 1.U,
+                                                                            (4.U) -> 2.U,
+                                                                            (8.U) -> 3.U
+                                                                        ))
     io.tlMasterTransmitter.bits.a_mask := io.reqIn.bits.activeByteLane
-    io.tlMasterTransmitter.bits.a_corrupt := 0.U
+    io.tlMasterTransmitter.bits.a_corrupt := false.B
     io.tlMasterTransmitter.valid := true.B
